+++ display/index.js
@@ -1,0 +1,3 @@
+const formatTokens = require('./format_tokens');
+
+module.exports = {formatTokens};