+++ peers/index.js
@@ -1,0 +1,3 @@
+const findKey = require('./find_key');
+
+module.exports = {findKey};